+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 ### Changes
 
 * Reduce `rubocop` `Layout/LineLength` metrics
+* Change typo in module name `OnlyofficeTelegramBugzillaNotifications`
+* Change typo in class name `TelegramBugzillaNotificaions`
 
 ## 0.1.0 (2020-08-18)